+++ Code/QStore/ISequenceSet.cs
@@ -9,8 +9,7 @@
         IEnumerable<T> GetByIndex(int index);
 
         IEnumerable<IEnumerable<T>> GetByPrefix(IEnumerable<T> prefix);
-
-        /// TODO: check if "int" is enough
-        int GetIndex(IEnumerable<T> sequence);
+        
+        long GetIndex(IEnumerable<T> sequence);
     }
 }
